Rename "redux" folder to "state"
