+++ client-src/state/categories/action-types.js
