+++ client-src/state/categories/initial-state.js
