+++ client-src/state/categories/reducer.js
