+++ client-src/transactions/components/content/index.js
@@ -2,8 +2,8 @@ import _ from 'lodash';
 import React, {Component} from 'react';
 import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
-import * as transactionsActionCreators from '../../../redux/transactions/action-creators';
-import * as categoriesActionCreators from '../../../redux/categories/action-creators';
+import * as transactionsActionCreators from '../../../state/transactions/action-creators';
+import * as categoriesActionCreators from '../../../state/categories/action-creators';
 import Subheader from '../subheader';
 import DateMenu from '../date-menu';
 import TransactionsList from '../transactions-list';
