+++ client-src/transactions/components/subheader/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
-import * as transactionsActionsCreators from '../../../redux/transactions/action-creators';
+import * as transactionsActionsCreators from '../../../state/transactions/action-creators';
 
 export function TransactionsSubheader({isOnline, transactionsActions}) {
   const disabled = !isOnline;
