+++ test/unit/client/state/categories/action-creators.js
@@ -1,9 +1,9 @@
 import xhr from 'xhr';
 import sinon from 'sinon';
-import * as actionCreators from '../../../../../client-src/redux/transactions/action-creators';
-import actionTypes from '../../../../../client-src/redux/transactions/action-types';
+import * as actionCreators from '../../../../../client-src/state/categories/action-creators';
+import actionTypes from '../../../../../client-src/state/categories/action-types';
 
-describe('transactions/actionCreators', function() {
+describe('categories/actionCreators', function() {
   beforeEach(() => {
     useFakeXMLHttpRequest();
     xhr.XMLHttpRequest = sinon.FakeXMLHttpRequest;
@@ -12,55 +12,54 @@ describe('transactions/actionCreators', function() {
     spy(xhr);
   });
 
-  describe('resetCreateTransactionResolution', () => {
+  describe('resetCreateCategoryResolution', () => {
     it('should return the right action', () => {
-      const result = actionCreators.resetCreateTransactionResolution();
+      const result = actionCreators.resetCreateCategoryResolution();
       expect(result).to.deep.equal({
-        type: actionTypes.CREATE_TRANSACTION_RESET_RESOLUTION
+        type: actionTypes.CREATE_CATEGORY_RESET_RESOLUTION
       });
     });
   });
 
-  describe('createTransaction', () => {
+  describe('createCategory', () => {
     it('should dispatch a begin action before making the request', () => {
-      const thunk = actionCreators.createTransaction({attributes: {label: 'pizza'}});
+      const thunk = actionCreators.createCategory({attributes: {label: 'pizza'}});
       thunk(this.dispatch);
       expect(this.dispatch).to.have.been.calledOnce;
       expect(this.dispatch).to.have.been.calledWithExactly({
-        type: actionTypes.CREATE_TRANSACTION,
+        type: actionTypes.CREATE_CATEGORY,
         resource: {
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
       expect(this.dispatch).to.have.been.calledBefore(xhr.post);
     });
 
     it('should generate the expected request', () => {
-      const thunk = actionCreators.createTransaction({attributes: {label: 'pizza'}});
+      const thunk = actionCreators.createCategory({attributes: {label: 'pizza'}});
       const req = thunk(this.dispatch);
       expect(req).to.be.instanceof(this.FakeXMLHttpRequest);
-      expect(req.url).to.equal('/api/transactions');
+      expect(req.url).to.equal('/api/categories');
       expect(req.method).to.equal('POST');
       const expectedBody = JSON.stringify({
         data: {
           attributes: {
             label: 'pizza'
           },
-          type: 'transactions',
+          type: 'categories',
         }
       });
       expect(req.requestBody).to.deep.equal(expectedBody);
     });
 
     it('should respond appropriately when there are no errors', () => {
-      const thunk = actionCreators.createTransaction({attributes: {label: 'pizza'}});
+      const thunk = actionCreators.createCategory({attributes: {label: 'pizza'}});
       const req = thunk(this.dispatch);
       const respBody = JSON.stringify({
         data: {
           id: 10,
+          type: 'categories',
           attributes: {
             label: 'pizza'
           }
@@ -69,100 +68,100 @@ describe('transactions/actionCreators', function() {
       req.respond(200, {}, respBody);
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.CREATE_TRANSACTION_SUCCESS,
-        transaction: {
+        type: actionTypes.CREATE_CATEGORY_SUCCESS,
+        resource: {
           id: 10,
+          type: 'categories',
           attributes: {
             label: 'pizza'
           }
         }
       });
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.CREATE_TRANSACTION_FAILURE
-      });
-    });
-
-    it('should respond appropriately when aborted', () => {
-      const thunk = actionCreators.createTransaction({attributes: {label: 'pizza'}});
-      const req = thunk(this.dispatch);
-      req.abort();
-      expect(this.dispatch).to.have.been.calledTwice;
-      expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.CREATE_TRANSACTION,
+        type: actionTypes.CREATE_CATEGORY_FAILURE,
         resource: {
-          type: 'transactions',
+          id: 10,
+          type: 'categories',
           attributes: {
             label: 'pizza'
           }
         }
       });
+    });
+
+    it('should respond appropriately when aborted', () => {
+      const thunk = actionCreators.createCategory({attributes: {label: 'pizza'}});
+      const req = thunk(this.dispatch);
+      req.abort();
+      expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.CREATE_TRANSACTION_ABORTED,
+        type: actionTypes.CREATE_CATEGORY,
         resource: {
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+          type: 'categories',
+          attributes: {label: 'pizza'}
+        }
+      });
+      expect(this.dispatch).to.have.been.calledWith({
+        type: actionTypes.CREATE_CATEGORY_ABORTED,
+        resource: {
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
     });
 
     it('should respond appropriately when a error status code is returned', () => {
-      const thunk = actionCreators.createTransaction({attributes: {label: 'pizza'}});
+      const thunk = actionCreators.createCategory({attributes: {label: 'pizza'}});
       const req = thunk(this.dispatch);
       req.respond(500);
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.CREATE_TRANSACTION_SUCCESS,
-        transaction: {
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+        type: actionTypes.CREATE_CATEGORY_SUCCESS,
+        resource: {
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.CREATE_TRANSACTION_FAILURE,
+        type: actionTypes.CREATE_CATEGORY_FAILURE,
         resource: {
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
     });
   });
 
-  describe('resetRetrieveTransactionsResolution', () => {
+  describe('resetRetrieveCategoriesResolution', () => {
     it('should return the right action', () => {
-      const result = actionCreators.resetRetrieveTransactionsResolution();
+      const result = actionCreators.resetRetrieveCategoriesResolution();
       expect(result).to.deep.equal({
-        type: actionTypes.RETRIEVE_TRANSACTIONS_RESET_RESOLUTION
+        type: actionTypes.RETRIEVE_CATEGORIES_RESET_RESOLUTION
       });
     });
   });
 
-  describe('retrieveTransactions', () => {
+  describe('retrieveCategories', () => {
     it('should dispatch a begin action before making the request', () => {
-      const thunk = actionCreators.retrieveTransactions();
+      const thunk = actionCreators.retrieveCategories();
       thunk(this.dispatch);
       expect(this.dispatch).to.have.been.calledOnce;
       expect(this.dispatch).to.have.been.calledWithExactly({
-        type: actionTypes.RETRIEVE_TRANSACTIONS
+        type: actionTypes.RETRIEVE_CATEGORIES
       });
       expect(this.dispatch).to.have.been.calledBefore(xhr.get);
     });
 
     it('should generate the expected request', () => {
-      const thunk = actionCreators.retrieveTransactions();
+      const thunk = actionCreators.retrieveCategories();
       const req = thunk(this.dispatch);
       expect(req).to.be.instanceof(this.FakeXMLHttpRequest);
-      expect(req.url).to.equal('/api/transactions');
+      expect(req.url).to.equal('/api/categories');
       expect(req.method).to.equal('GET');
     });
 
     it('should respond appropriately when there are no errors', () => {
-      const thunk = actionCreators.retrieveTransactions();
+      const thunk = actionCreators.retrieveCategories();
       const req = thunk(this.dispatch);
       const respBody = JSON.stringify({
         data: [
@@ -173,63 +172,63 @@ describe('transactions/actionCreators', function() {
       req.respond(200, {}, respBody);
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.RETRIEVE_TRANSACTIONS_SUCCESS,
-        transactions: [
+        type: actionTypes.RETRIEVE_CATEGORIES_SUCCESS,
+        resources: [
           {id: 1},
           {id: 2}
         ]
       });
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.RETRIEVE_TRANSACTIONS_FAILURE
+        type: actionTypes.RETRIEVE_CATEGORIES_FAILURE
       });
     });
 
     it('should respond appropriately when aborted', () => {
-      const thunk = actionCreators.retrieveTransactions();
+      const thunk = actionCreators.retrieveCategories();
       const req = thunk(this.dispatch);
       req.abort();
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.RETRIEVE_TRANSACTIONS
+        type: actionTypes.RETRIEVE_CATEGORIES
       });
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.RETRIEVE_TRANSACTIONS_ABORTED
+        type: actionTypes.RETRIEVE_CATEGORIES_ABORTED
       });
     });
 
     it('should respond appropriately when a error status code is returned', () => {
-      const thunk = actionCreators.retrieveTransactions();
+      const thunk = actionCreators.retrieveCategories();
       const req = thunk(this.dispatch);
       req.respond(500);
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.RETRIEVE_TRANSACTIONS_SUCCESS,
-        transactions: undefined
+        type: actionTypes.RETRIEVE_CATEGORIES_SUCCESS,
+        resources: undefined
       });
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.RETRIEVE_TRANSACTIONS_FAILURE
+        type: actionTypes.RETRIEVE_CATEGORIES_FAILURE
       });
     });
   });
 
-  describe('resetUpdateTransactionResolution', () => {
+  describe('resetUpdateCategoryResolution', () => {
     it('should return the right action', () => {
-      const result = actionCreators.resetUpdateTransactionResolution(2);
+      const result = actionCreators.resetUpdateCategoryResolution(2);
       expect(result).to.deep.equal({
-        type: actionTypes.UPDATE_TRANSACTION_RESET_RESOLUTION,
-        resourceId: 2
+        type: actionTypes.UPDATE_CATEGORY_RESET_RESOLUTION,
+        categoryId: 2
       });
     });
   });
 
-  describe('updateTransaction', () => {
+  describe('updateCategory', () => {
     beforeEach(() => {
       this.getState = function() {
         return {
-          transactions: {
-            transactions: [
-              {id: 2, type: 'transactions', attributes: {label: 'pizza'}},
-              {id: 10, type: 'transactions', attributes: {label: 'what'}}
+          categories: {
+            categories: [
+              {id: 2, type: 'categories', attributes: {label: 'pizza'}},
+              {id: 10, type: 'categories', attributes: {label: 'what'}}
             ]
           }
         };
@@ -241,14 +240,14 @@ describe('transactions/actionCreators', function() {
     });
 
     it('should dispatch a begin action before making the request', () => {
-      const thunk = actionCreators.updateTransaction({id: 10, attributes: {label: 'pizza'}});
+      const thunk = actionCreators.updateCategory({id: 10, attributes: {label: 'pizza'}});
       thunk(this.dispatch, this.getState);
       expect(this.dispatch).to.have.been.calledOnce;
       expect(this.dispatch).to.have.been.calledWithExactly({
-        type: actionTypes.UPDATE_TRANSACTION,
+        type: actionTypes.UPDATE_CATEGORY,
         resource: {
-          type: 'transactions',
           id: 10,
+          type: 'categories',
           attributes: {label: 'pizza'}
         }
       });
@@ -256,118 +255,99 @@ describe('transactions/actionCreators', function() {
     });
 
     it('should generate the expected request', () => {
-      const thunk = actionCreators.updateTransaction({id: 2, attributes: {label: 'pizza'}});
+      const thunk = actionCreators.updateCategory({id: 2, attributes: {label: 'pizza'}});
       const req = thunk(this.dispatch, this.getState);
       expect(req).to.be.instanceof(this.FakeXMLHttpRequest);
-      expect(req.url).to.equal('/api/transactions/2');
+      expect(req.url).to.equal('/api/categories/2');
       expect(req.method).to.equal('PATCH');
       const expectedBody = JSON.stringify({
         data: {
           id: 2,
-          attributes: {
-            label: 'pizza'
-          },
-          type: 'transactions',
+          attributes: {label: 'pizza'},
+          type: 'categories'
         }
       });
       expect(req.requestBody).to.deep.equal(expectedBody);
     });
 
     it('should respond appropriately when there are no errors', () => {
-      const thunk = actionCreators.updateTransaction({id: 2, attributes: {label: 'pizza'}});
+      const thunk = actionCreators.updateCategory({id: 2, attributes: {label: 'pizza'}});
       const req = thunk(this.dispatch, this.getState);
-      const respBody = JSON.stringify({
-        data: {
-          id: 2,
-          attributes: {
-            label: 'pizza'
-          },
-          type: 'transactions'
-        }
-      });
-      req.respond(200, {}, respBody);
+      req.respond(204, {});
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.UPDATE_TRANSACTION_SUCCESS,
+        type: actionTypes.UPDATE_CATEGORY_SUCCESS,
         resource: {
           id: 2,
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.UPDATE_TRANSACTION_FAILURE,
+        type: actionTypes.UPDATE_CATEGORY_FAILURE,
         resource: {
           id: 2,
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
     });
 
     it('should respond appropriately when aborted', () => {
-      const thunk = actionCreators.updateTransaction({id: 2, attributes: {label: 'pizza'}});
+      const thunk = actionCreators.updateCategory({id: 2, attributes: {label: 'pizza'}});
       const req = thunk(this.dispatch, this.getState);
       req.abort();
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.UPDATE_TRANSACTION,
+        type: actionTypes.UPDATE_CATEGORY,
         resource: {
-          type: 'transactions',
           id: 2,
+          type: 'categories',
           attributes: {label: 'pizza'}
         }
       });
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.UPDATE_TRANSACTION_ABORTED,
+        type: actionTypes.UPDATE_CATEGORY_ABORTED,
         resource: {
-          type: 'transactions',
           id: 2,
+          type: 'categories',
           attributes: {label: 'pizza'}
         }
       });
     });
 
     it('should respond appropriately when a error status code is returned', () => {
-      const thunk = actionCreators.updateTransaction({id: 2, attributes: {label: 'pizza'}});
+      const thunk = actionCreators.updateCategory({id: 2, attributes: {label: 'pizza'}});
       const req = thunk(this.dispatch, this.getState);
       req.respond(500);
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.UPDATE_TRANSACTION_SUCCESS,
+        type: actionTypes.UPDATE_CATEGORY_SUCCESS,
         resource: {
           id: 2,
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.UPDATE_TRANSACTION_FAILURE,
+        type: actionTypes.UPDATE_CATEGORY_FAILURE,
         resource: {
           id: 2,
-          type: 'transactions',
-          attributes: {
-            label: 'pizza'
-          }
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
     });
   });
 
-  describe('deleteTransaction', () => {
+  describe('deleteCategory', () => {
     beforeEach(() => {
       this.getState = function() {
         return {
-          transactions: {
-            transactions: [
-              {id: 2, type: 'transactions', attributes: {label: 'pizza'}},
-              {id: 10, type: 'transactions', attributes: {label: 'what'}}
+          categories: {
+            categories: [
+              {id: 2, type: 'categories', attributes: {label: 'pizza'}},
+              {id: 10, type: 'categories', attributes: {label: 'what'}}
             ]
           }
         };
@@ -379,68 +359,70 @@ describe('transactions/actionCreators', function() {
     });
 
     it('should dispatch a begin action before making the request', () => {
-      const thunk = actionCreators.deleteTransaction(2);
+      const thunk = actionCreators.deleteCategory(2);
       thunk(this.dispatch, this.getState);
       expect(this.dispatch).to.have.been.calledOnce;
       expect(this.dispatch).to.have.been.calledWithExactly({
-        type: actionTypes.DELETE_TRANSACTION,
+        type: actionTypes.DELETE_CATEGORY,
         resource: {
-          id: 2, type: 'transactions', attributes: {label: 'pizza'}
+          id: 2,
+          type: 'categories',
+          attributes: {label: 'pizza'}
         }
       });
       expect(this.dispatch).to.have.been.calledBefore(xhr.del);
     });
 
     it('should generate the expected request', () => {
-      const thunk = actionCreators.deleteTransaction(2);
+      const thunk = actionCreators.deleteCategory(2);
       const req = thunk(this.dispatch, this.getState);
       expect(req).to.be.instanceof(this.FakeXMLHttpRequest);
-      expect(req.url).to.equal('/api/transactions/2');
+      expect(req.url).to.equal('/api/categories/2');
       expect(req.method).to.equal('DELETE');
     });
 
     it('should respond appropriately when there are no errors', () => {
-      const thunk = actionCreators.deleteTransaction(2);
+      const thunk = actionCreators.deleteCategory(2);
       const req = thunk(this.dispatch, this.getState);
       req.respond(200);
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.DELETE_TRANSACTION_SUCCESS,
-        resource: {id: 2, type: 'transactions', attributes: {label: 'pizza'}}
+        type: actionTypes.DELETE_CATEGORY_SUCCESS,
+        resource: {id: 2, type: 'categories', attributes: {label: 'pizza'}}
       });
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.DELETE_TRANSACTION_FAILURE,
-        resource: {id: 2, type: 'transactions', attributes: {label: 'pizza'}}
+        type: actionTypes.DELETE_CATEGORY_FAILURE,
+        resource: {id: 2, type: 'categories', attributes: {label: 'pizza'}}
       });
     });
 
     it('should respond appropriately when aborted', () => {
-      const thunk = actionCreators.deleteTransaction(2);
+      const thunk = actionCreators.deleteCategory(2);
       const req = thunk(this.dispatch, this.getState);
       req.abort();
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.DELETE_TRANSACTION,
-        resource: {id: 2, type: 'transactions', attributes: {label: 'pizza'}}
+        type: actionTypes.DELETE_CATEGORY,
+        resource: {id: 2, type: 'categories', attributes: {label: 'pizza'}}
       });
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.DELETE_TRANSACTION_ABORTED,
-        resource: {id: 2, type: 'transactions', attributes: {label: 'pizza'}}
+        type: actionTypes.DELETE_CATEGORY_ABORTED,
+        resource: {id: 2, type: 'categories', attributes: {label: 'pizza'}}
       });
     });
 
     it('should respond appropriately when a error status code is returned', () => {
-      const thunk = actionCreators.deleteTransaction(2);
+      const thunk = actionCreators.deleteCategory(2);
       const req = thunk(this.dispatch, this.getState);
       req.respond(500);
       expect(this.dispatch).to.have.been.calledTwice;
       expect(this.dispatch).to.not.have.been.calledWith({
-        type: actionTypes.DELETE_TRANSACTION_SUCCESS,
-        resource: {id: 2, type: 'transactions', attributes: {label: 'pizza'}}
+        type: actionTypes.DELETE_CATEGORY_SUCCESS,
+        resource: {id: 2, type: 'categories', attributes: {label: 'pizza'}}
       });
       expect(this.dispatch).to.have.been.calledWith({
-        type: actionTypes.DELETE_TRANSACTION_FAILURE,
-        resource: {id: 2, type: 'transactions', attributes: {label: 'pizza'}}
+        type: actionTypes.DELETE_CATEGORY_FAILURE,
+        resource: {id: 2, type: 'categories', attributes: {label: 'pizza'}}
       });
     });
   });
